Refactor: AppLayout 안에 페이지 구현하도록 변경
