+++ client/src/main.jsx
@@ -1,20 +1,18 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { Outlet, RouterProvider, createBrowserRouter } from 'react-router-dom';
-import './index.css';
+import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { AuthProvider } from './contexts/AuthContext.jsx';
-// import App from './App.jsx';
+import AppLayout from './components/ui/AppLayout.jsx';
 import MakeCommunity from './pages/MakeCommunityPage.jsx';
 import ChooseFeature from './pages/ChooseFeaturePage.jsx';
 import NamingCommunity from './pages/NamingCommunity.jsx';
 import Home from './pages/Home.jsx';
 import ChooseInterest from './pages/ChooseInterest.jsx';
-import Login from './pages/loginPage.jsx';
 import FindLocation from './pages/FindLocationPage.jsx';
 import Detail from './pages/Detail.jsx';
 import SearchPage from './pages/Search.jsx';
 import Alarm from './pages/Alarm.jsx';
 import CategoryPage from './pages/Category.jsx';
+import Login from './pages/loginPage.jsx';
 import LoginPage from './pages/LocalLogin/LoginPage.jsx';
 import LinkSample from './pages/LocalLogin/LinkSample.jsx';
 import ProtectedRoute from './pages/ProtectedRoute.jsx';
@@ -41,7 +39,7 @@ const router = createBrowserRouter([
     element: (
       <AuthProvider>
         <ProtectedRoute>
-          <Outlet />
+          <AppLayout />
         </ProtectedRoute>
       </AuthProvider>
     ),
